requirements.txt
discord
asyncio
mysql
python-dotenv
smtplib